all: Extract FuzzedDataProviderImpl into the driver
