+++ agent/src/test/java/com/code_intelligence/jazzer/runtime/BUILD.bazel
@@ -1,30 +1,6 @@
 load("//bazel:compat.bzl", "SKIP_ON_WINDOWS")
 
 java_test(
-    name = "FuzzedDataProviderImplTest",
-    srcs = ["FuzzedDataProviderImplTest.java"],
-    use_testrunner = False,
-    deps = [
-        "//agent/src/main/java/com/code_intelligence/jazzer/api",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:fuzzed_data_provider",
-        "//driver/src/main/native/com/code_intelligence/jazzer/driver:jazzer_driver",
-    ],
-)
-
-java_test(
-    name = "RecordingFuzzedDataProviderTest",
-    srcs = [
-        "RecordingFuzzedDataProviderTest.java",
-    ],
-    deps = [
-        "//agent/src/main/java/com/code_intelligence/jazzer/api",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:fuzzed_data_provider",
-        "@maven//:junit_junit",
-    ],
-)
-
-java_test(
     name = "TraceCmpHooksTest",
     srcs = [
         "TraceCmpHooksTest.java",
