+++ driver/src/main/java/com/code_intelligence/jazzer/driver/BUILD.bazel
@@ -27,7 +27,9 @@ java_jni_library(
         "//driver/src/test:__subpackages__",
     ],
     deps = [
+        ":fuzzed_data_provider_impl",
         ":opt",
+        ":recording_fuzzed_data_provider",
         ":reproducer_template",
         ":signal_handler",
         ":utils",
@@ -36,9 +38,19 @@ java_jni_library(
         "//agent/src/main/java/com/code_intelligence/jazzer/instrumentor",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:coverage_map",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:fuzzed_data_provider",
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime:unsafe_provider",
         "//agent/src/main/java/com/code_intelligence/jazzer/utils",
+    ],
+)
+
+java_jni_library(
+    name = "fuzzed_data_provider_impl",
+    srcs = ["FuzzedDataProviderImpl.java"],
+    native_libs = ["//driver/src/main/native/com/code_intelligence/jazzer/driver:jazzer_fuzzed_data_provider"],
+    visibility = ["//driver:__subpackages__"],
+    deps = [
+        "//agent/src/main/java/com/code_intelligence/jazzer/api",
+        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:unsafe_provider",
     ],
 )
 
@@ -56,6 +68,13 @@ java_library(
         "//agent/src/main/java/com/code_intelligence/jazzer:__subpackages__",
         "//driver/src/test/java/com/code_intelligence/jazzer/driver:__pkg__",
     ],
+)
+
+java_library(
+    name = "recording_fuzzed_data_provider",
+    srcs = ["RecordingFuzzedDataProvider.java"],
+    visibility = ["//driver/src/test/java/com/code_intelligence/jazzer/driver:__pkg__"],
+    deps = ["//agent/src/main/java/com/code_intelligence/jazzer/api"],
 )
 
 java_jni_library(
