+++ driver/src/main/java/com/code_intelligence/jazzer/driver/FuzzTargetRunner.java
@@ -24,9 +24,7 @@ import com.code_intelligence.jazzer.api.FuzzedDataProvider;
 import com.code_intelligence.jazzer.autofuzz.FuzzTarget;
 import com.code_intelligence.jazzer.instrumentor.CoverageRecorder;
 import com.code_intelligence.jazzer.runtime.CoverageMap;
-import com.code_intelligence.jazzer.runtime.FuzzedDataProviderImpl;
 import com.code_intelligence.jazzer.runtime.JazzerInternal;
-import com.code_intelligence.jazzer.runtime.RecordingFuzzedDataProvider;
 import com.code_intelligence.jazzer.runtime.UnsafeProvider;
 import com.code_intelligence.jazzer.utils.ExceptionUtils;
 import com.code_intelligence.jazzer.utils.ManifestUtils;
