+++ driver/src/main/java/com/code_intelligence/jazzer/replay/BUILD.bazel
@@ -3,10 +3,9 @@ load("@fmeum_rules_jni//jni:defs.bzl", "java_jni_library")
 java_jni_library(
     name = "replay",
     srcs = ["Replayer.java"],
-    native_libs = ["//driver/src/main/native/com/code_intelligence/jazzer/driver:fuzzed_data_provider_standalone"],
     deps = [
         "//agent/src/main/java/com/code_intelligence/jazzer/api",
-        "//agent/src/main/java/com/code_intelligence/jazzer/runtime:fuzzed_data_provider",
+        "//driver/src/main/java/com/code_intelligence/jazzer/driver:fuzzed_data_provider_impl",
     ],
 )
