+++ driver/src/test/java/com/code_intelligence/jazzer/driver/FuzzedDataProviderImplTest.java
@@ -12,9 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.code_intelligence.jazzer.runtime;
+package com.code_intelligence.jazzer.driver;
 
 import com.code_intelligence.jazzer.api.FuzzedDataProvider;
+import com.code_intelligence.jazzer.driver.FuzzedDataProviderImpl;
 import java.util.Arrays;
 import java.util.stream.Collectors;
