+++ driver/src/test/java/com/code_intelligence/jazzer/driver/RecordingFuzzedDataProviderTest.java
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.code_intelligence.jazzer.runtime;
+package com.code_intelligence.jazzer.driver;
 
 import com.code_intelligence.jazzer.api.CannedFuzzedDataProvider;
 import com.code_intelligence.jazzer.api.FuzzedDataProvider;
+import com.code_intelligence.jazzer.driver.RecordingFuzzedDataProvider;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.stream.Collectors;
