+++ examples/BUILD.bazel
@@ -219,7 +219,7 @@ java_binary(
     main_class = "com.code_intelligence.jazzer.replay.Replayer",
     runtime_deps = [
         ":JsonSanitizerDenylistFuzzer_target_deploy.jar",
-        "//agent/src/main/java/com/code_intelligence/jazzer/replay:Replayer_deploy.jar",
+        "//driver/src/main/java/com/code_intelligence/jazzer/replay:Replayer_deploy.jar",
     ],
 )
